[Util] Move sim util helpers into one file and update imports
