+++ src/finn/util/fpgadataflow.py
@@ -29,8 +29,17 @@
 import os
 import warnings
 from qonnx.core.datatype import DataType
+from qonnx.core.modelwrapper import ModelWrapper
 from qonnx.custom_op.registry import getCustomOp, is_custom_op
 from qonnx.util.basic import get_by_name
+
+
+def is_mlo(model: ModelWrapper) -> bool:
+    """Returns True if the model is an MLO model (contains FINNLoop), False otherwise."""
+    for node in model.graph.node:
+        if node.op_type == "FINNLoop":
+            return True
+    return False
 
 
 def is_fpgadataflow_node(node):
